new claimsat/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
Shapely==2.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
geopy==2.4.1
Levenshtein==0.23.0
python-dotenv==1.0.0
aiofiles==23.2.1